backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
together==1.3.4
firebase-admin==6.6.0
gtts==2.5.4
ffmpeg-python==0.2.0
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.20
aiofiles==24.1.0
cloudinary==1.44.1
requests==2.31.0
Pillow==10.4.0
PyJWT==2.8.0
email-validator==2.1.0
